Support pydantic fields with alias in `query_to_dict` function
Sometimes QueryDict parameters will have invalid characters that cannot be used as
modelfield names. Pydantic supports `alias` for this on its Field class.

pyngo doesn't check for an alias in the modelfield, so if you have an
alias set, the returned data is invalid if the data is supposed to be of
a sequence type, as it only uses `get` on the QueryDict, even though the
model expects a list.

The added test should highlight the issue this patch addresses